feat: added average metric
