+++ src/controllers/server.controller.ts
@@ -1,5 +1,4 @@
 import {authenticate} from '@loopback/authentication';
-import {AnyObject, repository} from '@loopback/repository';
 import {
   post,
   get,
@@ -7,24 +6,18 @@ import {
   patch,
   requestBody,
   response,
-  HttpErrors,
   param,
 } from '@loopback/rest';
-import {config} from '../config';
 import {PermissionKeys} from '../enums';
 import {Server} from '../models';
-import {ServerRepository, UserRepository} from '../repositories';
-import {PolkadotJs} from '../utils/polkadotJs-utils';
-
-const {getKeyring} = new PolkadotJs();
+import {ServerService} from '../services';
+import {service} from '@loopback/core';
 
 @authenticate({strategy: 'jwt', options: {required: [PermissionKeys.ADMIN]}})
 export class ServerController {
   constructor(
-    @repository(ServerRepository)
-    public serverRepository: ServerRepository,
-    @repository(UserRepository)
-    public userRepository: UserRepository,
+    @service(ServerService)
+    public serverService: ServerService,
   ) {}
 
   @authenticate.skip()
@@ -40,66 +33,11 @@ export class ServerController {
       },
     },
   })
-  async find(@param.query.boolean('median') median?: boolean): Promise<Server> {
-    const server = await this.getServer();
-    if (!server) throw new HttpErrors.NotFound('ServerNotFound');
-
-    if (median) {
-      const [
-        userPost,
-        userComment,
-        userExperience,
-        userTransaction,
-        userSubscription,
-      ] = await Promise.all([
-        this.userRepository.find({
-          order: ['metric.totalPosts ASC'],
-          skip: Math.ceil(server.metric.totalPosts.totalAll / 2),
-          limit: 1,
-        }),
-        this.userRepository.find({
-          order: ['metric.totalComments ASC'],
-          skip: Math.ceil(server.metric.totalComments / 2),
-          limit: 1,
-        }),
-        this.userRepository.find({
-          order: ['metric.totalExperiences ASC'],
-          skip: Math.ceil(server.metric.totalExperiences / 2),
-          limit: 1,
-        }),
-        this.userRepository.find({
-          order: ['metric.totalTransactions ASC'],
-          skip: Math.ceil(server.metric.totalTransactions / 2),
-          limit: 1,
-        }),
-        this.userRepository.find({
-          order: ['metric.totalSubscriptions ASC'],
-          skip: Math.ceil(server.metric.totalSubscriptions / 2),
-          limit: 1,
-        }),
-      ]);
-
-      const medianPost = userPost?.[0]?.metric?.totalPosts ?? 0;
-      const medianComment = userComment?.[0]?.metric?.totalComments ?? 0;
-      const medianExperience =
-        userExperience?.[0]?.metric?.totalExperiences ?? 0;
-      const medianTransaction =
-        userTransaction?.[0]?.metric?.totalTransactions ?? 0;
-      const medianSubscription =
-        userSubscription?.[0]?.metric?.totalSubscriptions ?? 0;
-
-      Object.assign(server, {
-        median: {
-          medianPost,
-          medianComment,
-          medianExperience,
-          medianTransaction,
-          medianSubscription,
-        },
-      });
-    }
-
-    return server;
+  async find(
+    @param.query.boolean('median') median?: boolean,
+    @param.query.boolean('average') average?: boolean,
+  ): Promise<Server> {
+    return this.serverService.find(median, average);
   }
 
   @post('/server')
@@ -120,13 +58,7 @@ export class ServerController {
     })
     server: Omit<Server, 'id'>,
   ): Promise<Server> {
-    const exists = await this.getServer();
-
-    if (exists) {
-      throw new HttpErrors.UnprocessableEntity('ServerExist');
-    }
-
-    return this.serverRepository.create(server);
+    return this.serverService.create(server);
   }
 
   @patch('/server')
@@ -139,41 +71,13 @@ export class ServerController {
         'application/json': {
           schema: getModelSchemaRef(Server, {
             partial: true,
-            exclude: ['id', 'metric', 'median'],
+            exclude: ['id', 'metric', 'median', 'average'],
           }),
         },
       },
     })
     server: Partial<Server>,
   ): Promise<void> {
-    const currentServer = await this.getServer();
-
-    if (!currentServer) {
-      throw new HttpErrors.UnprocessableEntity('ServerNotExist');
-    }
-
-    const serverId = currentServer.id;
-
-    if (server.accountId || server.images) {
-      server.accountId = {...currentServer.accountId, ...server.accountId};
-      server.images = {...currentServer.images, ...server.images};
-    }
-
-    return this.serverRepository.updateById(serverId, server);
-  }
-
-  async getServer(): Promise<Server | null> {
-    const mnemonic = config.MYRIAD_ADMIN_MNEMONIC;
-
-    if (!mnemonic) throw new HttpErrors.NotFound('MnemonicNotFound');
-
-    const serverAdmin = getKeyring().addFromMnemonic(mnemonic);
-    const address = serverAdmin.address;
-
-    return this.serverRepository.findOne(<AnyObject>{
-      where: {
-        'accountId.myriad': address,
-      },
-    });
+    return this.serverService.update(server);
   }
 }
